mvc + logger
