--- src/main/java/ua/edu/sumdu/j2se/shestopalova/tasks/TaskListFactory.java
@@ -1,15 +1,0 @@
-package ua.edu.sumdu.j2se.shestopalova.tasks;
-
-public class TaskListFactory {
-
-    public static AbstractTaskList createTaskList(ListTypes.types types) {
-        switch (types) {
-            case LINKED:
-                return new LinkedTaskList();
-            case ARRAY:
-            default:
-                return new ArrayTaskList();
-        }
-
-    }
-}
