+++ src/main/java/ua/edu/sumdu/j2se/shestopalova/tasks/controller/TaskController.java
@@ -1,0 +1,18 @@
+package ua.edu.sumdu.j2se.shestopalova.tasks.controller;
+
+public interface TaskController {
+
+    void addTask();
+
+
+    void editTask();
+
+
+    void removeTask();
+
+
+    void showTaskList();
+
+
+    void showCalendar();
+}
