+++ src/main/java/ua/edu/sumdu/j2se/shestopalova/tasks/model/AbstractTaskList.java
@@ -1,6 +1,6 @@
-package ua.edu.sumdu.j2se.shestopalova.tasks;
+package ua.edu.sumdu.j2se.shestopalova.tasks.model;
+
 import java.util.Iterator;
-import java.util.Objects;
 import java.util.stream.Stream;
 import java.io.Serializable;
