+++ src/main/java/ua/edu/sumdu/j2se/shestopalova/tasks/model/ListTypes.java
@@ -1,4 +1,4 @@
-package ua.edu.sumdu.j2se.shestopalova.tasks;
+package ua.edu.sumdu.j2se.shestopalova.tasks.model;
 
 public class ListTypes {
     public enum types {
