+++ src/main/java/ua/edu/sumdu/j2se/shestopalova/tasks/model/Tasks.java
@@ -1,4 +1,4 @@
-package ua.edu.sumdu.j2se.shestopalova.tasks;
+package ua.edu.sumdu.j2se.shestopalova.tasks.model;
 
 import java.time.LocalDateTime;
 import java.util.*;
