+++ test.json
@@ -1,1 +1,1 @@
-{"taskArray":[{"time":{"date":{"year":2022,"month":7,"day":5},"time":{"hour":0,"minute":0,"second":0,"nano":585545200}},"title":"A","active":false,"repeated":false,"interval":0},{"title":"B","active":false,"repeated":true,"start_time":{"date":{"year":2022,"month":7,"day":5},"time":{"hour":0,"minute":0,"second":0,"nano":585545200}},"end_time":{"date":{"year":2022,"month":7,"day":6},"time":{"hour":0,"minute":0,"second":0,"nano":585545200}},"interval":3600},{"time":{"date":{"year":2022,"month":7,"day":5},"time":{"hour":0,"minute":0,"second":0,"nano":585545200}},"title":"C","active":true,"repeated":false,"interval":0},{"title":"B","active":true,"repeated":true,"start_time":{"date":{"year":2022,"month":7,"day":5},"time":{"hour":0,"minute":0,"second":0,"nano":585545200}},"end_time":{"date":{"year":2022,"month":7,"day":6},"time":{"hour":0,"minute":0,"second":0,"nano":585545200}},"interval":3600},null,null,null,null,null,null],"size":4}
+{"taskArray":[{"time":{"date":{"year":2022,"month":7,"day":19},"time":{"hour":0,"minute":0,"second":0,"nano":494405200}},"title":"A","active":false,"repeated":false,"interval":0},{"title":"B","active":false,"repeated":true,"start_time":{"date":{"year":2022,"month":7,"day":19},"time":{"hour":0,"minute":0,"second":0,"nano":494405200}},"end_time":{"date":{"year":2022,"month":7,"day":20},"time":{"hour":0,"minute":0,"second":0,"nano":494405200}},"interval":3600},{"time":{"date":{"year":2022,"month":7,"day":19},"time":{"hour":0,"minute":0,"second":0,"nano":494405200}},"title":"C","active":true,"repeated":false,"interval":0},{"title":"B","active":true,"repeated":true,"start_time":{"date":{"year":2022,"month":7,"day":19},"time":{"hour":0,"minute":0,"second":0,"nano":494405200}},"end_time":{"date":{"year":2022,"month":7,"day":20},"time":{"hour":0,"minute":0,"second":0,"nano":494405200}},"interval":3600},null,null,null,null,null,null],"size":4}
